// src/interFace/interFace.ts
export interface IBlogDataType {
  id: number;
  img: StaticImageData;
  blogTag?: string;
  author: string;
  calender: string;
  detail: string;
  description?: string;
}